landing done
finished stylizing landing, moved landing to pages, added images (landing), added button component, not finished (TODO)
