+++ src/components/Button/Button.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import s from "./Button.module.css";
+
+const Button = () => {
+  return (
+    <div>
+      <button type="button" className={s.button}></button>
+    </div>
+  );
+};
+
+export default Button;
