+++ src/components/Button/Button.module.css
@@ -1,0 +1,10 @@
+.button {
+  height: 52px;
+  width: 52px;
+  color: #ffffff;
+  background-color: #ff851c;
+  border-radius: 50%;
+  border-color: #ff851c;
+  border-style: none;
+  box-shadow: 0px 0px 4px 2px #ff851c;
+}
