+++ src/pages/Landing/Landing.module.css
@@ -1,9 +1,19 @@
+.landing {
+  width: 100%;
+  height: 650px;
+  background: url(/images/Landing/Mobile/landingMB.png) no-repeat 0% 100%,
+    url(/images/Landing/Mobile/landingMG.png) no-repeat 0% 100%;
+}
+
+.position {
+  padding-top: 60px;
+}
+
 .title {
   color: #00d7ff;
   font-family: "HelveticaNeueCyr-bold";
   font-size: 24px;
   padding-bottom: 70px;
-  padding-top: 60px;
   font-weight: 700;
   font-style: bold;
 }
@@ -24,7 +34,6 @@
   font-size: 14px;
   font-weight: 400;
   font-style: normal;
-  padding-bottom: 30px;
 }
 
 .spacer {
@@ -53,6 +62,14 @@
 }
 
 @media screen and (min-width: 530px) {
+  .landing {
+    height: 1020px;
+    background: url(/images/Landing/Tablet/landingT1.png) no-repeat 100% 0%,
+      url(/images/Landing/Tablet/landingT2.png) no-repeat 100% 100%;
+  }
+  .position {
+    padding-top: 310px;
+  }
   .title {
     font-size: 20px;
   }
@@ -61,7 +78,7 @@
     height: 100px;
     font-size: 24px;
   }
-  .form-desc {
+  .form__desc {
     font-size: 18px;
   }
   .input {
@@ -70,9 +87,21 @@
 }
 
 @media screen and (min-width: 1024px) {
+  .landing {
+    height: 850px;
+    background: url(/images/Landing/Desktop/landingD1.png) no-repeat 100% 0%,
+      url(/images/Landing/Desktop/landingD2.png) no-repeat 100% 100%;
+  }
+  .position {
+    padding-top: 195px;
+    padding-left: 160px;
+  }
   .about {
     font-size: 34px;
     width: 560px;
     height: 135px;
+  }
+  .form__desc {
+    font-size: 24px;
   }
 }
